Add macCatalyst to the PlatformChecks (#524)
### Motivation

When attempting to compile the Swift OpenAPI Generator on MacCatalyst,
code that compiles on iOS will _not_ compile as it will fail on the
`PlatformChecks`.

### Modifications

Added an extra condition in `PlatformChecks` to allow running on
MacCatalyst.

### Result

fixes #523

### Test Plan

I've tried to familiarize myself with the project, but could not come up
with a good way to add a test for running the tool on MacCatalyst. In my
manual testing giving the instructions in the issue, it now works as
expected.