[GHA] Call the reusable package workflow (#588)
# Motivation

We are migrating our CI over to GH actions. We have created a reusable
workflow inside the NIO repository that we want to move into a separate
repo in the future. For testing purposes let's try out calling the
reusable workflow from the this repo as well.

# Modification

Add a GH actions workflow that calls the reusable package workflow.