ci: Remove Docker Compose files used by old CI (#633)
### Motivation

In #604 we enabled all the Github Actions based workflows we need for
CI. Since then we have updated the branch protection rules and disabled
the webhook for the old CI, so we can remove the Docker Compose files
that were used by the old CI.

### Modifications

Delete the Docker Compose files.

### Result

Less clutter in the repo.

### Test Plan

This PR should show the new, Github Actions CI checks, and shouldn't be
showing any of the older CI jobs. If that's true, we are safe to remove
them.