Fix command plugin when run on targets with dependencies (#801)
### Motivation

The command plugin is used for ahead-of-time (AOT) code generation. It
can be run with `--target`, or without, in which case it considers all
targets based on the presence of the requisite OpenAPI document and
config file.

If a target is detected that doesn't have either of these files, it
should be considered an error if that target was explicitly asked for
using `--target`. If it wasn't—either because we're looking at all
targets, or because it is a dependency—it should be skipped.

### Modifications

- Extend integration test to repro issue for AOT targets with
dependencies
- Skip targets that have no requisite files for generation unless they
were asked for

### Result

Fixes command plugin when used for AOT generation with targets with
target dependencies.

### Test Plan

This PR starts out with a commit that updated the integration test to
repro the issue, which we should see fail in CI. Only then was a commit
added to address the issue, restoring the CI to green.