Allow recursive schemas in multipart-based paths (#813)
### Motivation

Fixes #769
Currently, as seen in the issue mentioned above, if a recursive schema
gets referenced from a multipart-based path, a reference cycle error
gets thrown. Since we support recursive type generation already, this
shouldn't happen.

### Modifications

This PR removes the last `dereferenced(in:)` method call, which would
throw if a reference cycle was found, replacing it with a simple lookup
in the OpenAPI components tree. It also replaces
`DerereferencedJSONSchema` with `JSONSchema` accordingly.

### Result

Multipart-based paths can now reference recursive types.

### Test Plan

This also adds a test with such a case. We could probably also add a
test with a multipart path referencing an _array_ of references instead
of just a reference to be thorough.