[Docs] Document Android support (#852)
### Motivation

See https://github.com/apple/swift-openapi-runtime/pull/163.

### Modifications

Updated the supported platforms in docs to include Android.

Also removed tracking the last Swift version, as we haven't been keeping
that up to date, so it was stale.

### Result

Docs representing that Swift OpenAPI Generator can be used for Android
projects.

### Test Plan

N/A